+++ Bridge/PortalNode/Configuration/PortalNodeConfigurationHelper.php
@@ -51,7 +51,14 @@ class PortalNodeConfigurationHelper
     public function json(string $jsonFile, ?array $mappings = null): \Closure
     {
         return function () use ($mappings, $jsonFile): array {
-            $config = \json_decode(\file_get_contents($jsonFile), true, 512, \JSON_THROW_ON_ERROR);
+            $json = \file_get_contents($jsonFile);
+
+            if (!\is_string($json)) {
+                throw new \RuntimeException('Can not read JSON file', 1647801830);
+            }
+
+
+            $config = \json_decode($json, true, 512, \JSON_THROW_ON_ERROR);
 
             if (!\is_array($config)) {
                 throw new \RuntimeException('Can not load JSON file', 1647801829);